Pull extensions out to their own files.
